feat(P1-003): Move Client and JWTSession classes to browser-common
- Move Client.ts and JWTSession.ts from js/src/ to browser-common/src/
- Move supporting utilities (storage.ts, CloseEvent.ts) to browser-common/src/utils/
- Move related test files (JWTSession.test.ts, storage.test.ts) to browser-common
- Update all import references across the codebase to use @signalwire/browser-common
- Add jwt-decode dependency to browser-common package.json
- Update browser-common exports to include Client, JWTSession, and utilities

This migration enables sharing of client infrastructure between Video SDK and
Call Fabric SDK as outlined in the migration plan. All tests pass confirming:
- Client initialization works correctly
- JWT session management functions properly
- WebSocket connections establish successfully

Files moved:
- packages/js/src/Client.ts → packages/browser-common/src/Client.ts
- packages/js/src/JWTSession.ts → packages/browser-common/src/JWTSession.ts
- packages/js/src/utils/storage.ts → packages/browser-common/src/utils/storage.ts
- packages/js/src/utils/CloseEvent.ts → packages/browser-common/src/utils/CloseEvent.ts
- packages/js/src/JWTSession.test.ts → packages/browser-common/src/JWTSession.test.ts
- packages/js/src/utils/storage.test.ts → packages/browser-common/src/utils/storage.test.ts

Updated imports in:
- packages/js/src/createClient.ts
- packages/js/src/fabric/SATSession.ts
- packages/js/src/testUtils.ts
- packages/js/src/createRoomObject.ts
- packages/js/src/video/RoomSession.ts
- packages/js/src/video/index.ts
- packages/js/src/fabric/FabricRoomSession.ts
- packages/js/src/fabric/WSClient.ts
