+++ packages/browser-common/package.json
@@ -15,7 +15,8 @@
   },
   "dependencies": {
     "@signalwire/core": "4.3.0-dev.202506111650.ad2f5be.0",
-    "@signalwire/webrtc": "3.14.0-dev.202506111650.ad2f5be.0"
+    "@signalwire/webrtc": "3.14.0-dev.202506111650.ad2f5be.0",
+    "jwt-decode": "^3.1.2"
   },
   "devDependencies": {
     "@types/jest": "^29.5.2",
