+++ packages/browser-common/src/JWTSession.ts
