+++ packages/browser-common/src/index.ts
@@ -6,8 +6,28 @@
  * the Video SDK and the Call Fabric SDK.
  */
 
-// Placeholder for browser-common exports
-// This will be populated with actual shared components during the migration
+// BaseRoomSession - Core session management
+export {
+  BaseRoomSession,
+  BaseRoomSessionConnection,
+  BaseRoomSessionOptions,
+  BaseRoomSessionEvents,
+  createBaseRoomSessionObject,
+} from './BaseRoomSession'
+
+// Client - WebSocket client for SignalWire
+export {
+  Client,
+  ClientAPI,
+  MakeRoomOptions,
+} from './Client'
+
+// JWTSession - JWT session management for browser
+export { JWTSession } from './JWTSession'
+
+// Utilities
+export { getStorage, sessionStorageManager } from './utils/storage'
+export { SwCloseEvent } from './utils/CloseEvent'
 
 export const BROWSER_COMMON_VERSION = '0.0.0'
