+++ packages/browser-common/src/utils/CloseEvent.ts
