+++ packages/browser-common/src/utils/storage.ts
