+++ packages/js/tsconfig.build.json
@@ -4,6 +4,6 @@
     "emitDeclarationOnly": true,
     "outDir": "dist"
   },
-  "include": ["./src/**/*.ts", "../core/src/**/*.ts", "../webrtc/src/**/*.ts"],
+  "include": ["./src/**/*.ts", "../core/src/**/*.ts", "../webrtc/src/**/*.ts", "../browser-common/src/**/*.ts"],
   "exclude": ["**/*.test.ts"]
 }
